requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic-settings
httpx
phonenumbers
pypdf
pdf2image
pillow
pytesseract
python-multipart
aiofiles
pytest
pytest-asyncio
httpx
